+++ src/main/java/coffee/synyx/auth/web/LoginController.java
@@ -8,11 +8,19 @@ import org.springframework.beans.factory.annotation.Autowired;
 
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 
+import org.springframework.cloud.client.ServiceInstance;
+import org.springframework.cloud.client.discovery.DiscoveryClient;
+import org.springframework.cloud.netflix.eureka.EurekaDiscoveryClient.EurekaServiceInstance;
+
 import org.springframework.stereotype.Controller;
+
+import org.springframework.ui.Model;
 
 import org.springframework.web.bind.annotation.RequestMapping;
 
 import java.security.Principal;
+
+import java.util.List;
 
 import static org.slf4j.LoggerFactory.getLogger;
 
@@ -31,19 +39,23 @@ public class LoginController {
     private static final Logger LOGGER = getLogger(lookup().lookupClass());
 
     private final AuthConfigurationProperties authConfigurationProperties;
+    private final DiscoveryClient discoveryClient;
 
     @Autowired
-    public LoginController(AuthConfigurationProperties authConfigurationProperties) {
+    public LoginController(AuthConfigurationProperties authConfigurationProperties, DiscoveryClient discoveryClient) {
 
         this.authConfigurationProperties = authConfigurationProperties;
+        this.discoveryClient = discoveryClient;
     }
 
     @RequestMapping(value = "/login", method = GET)
-    public String getLoginView(Principal principal) {
+    public String getLoginView(Model model, Principal principal) {
 
         String view = "auth/login";
 
-        if (principal != null) {
+        if (principal == null) {
+            model.addAttribute("passwordRecoveryUrl", getPasswordRecoveryUri());
+        } else {
             view = "redirect:" + authConfigurationProperties.getDefaultRedirectUrl();
         }
 
@@ -51,5 +63,25 @@ public class LoginController {
             view);
 
         return view;
+    }
+
+
+    private String getPasswordRecoveryUri() {
+
+        String passwordRecoveryServiceName = authConfigurationProperties.getPasswordRecoveryServiceName();
+        List<ServiceInstance> profileServices = discoveryClient.getInstances(passwordRecoveryServiceName);
+
+        if (!profileServices.isEmpty()) {
+            ServiceInstance profileService = profileServices.get(0);
+
+            if (profileService instanceof EurekaServiceInstance) {
+                EurekaServiceInstance eurekaServiceProfile = (EurekaServiceInstance) profileService;
+
+                return eurekaServiceProfile.getInstanceInfo().getHomePageUrl()
+                    + authConfigurationProperties.getPasswordRecoveryPath();
+            }
+        }
+
+        return null;
     }
 }